Refactors ssr tests
