--- .babelrc
@@ -1,7 +1,0 @@
-{
-  "presets": [
-    [
-      "@babel/preset-env"
-    ]
-  ]
-}
